requirements/base.txt
# Django
# ------------------------------------------------
# >=1.11.9,<2.0
django==2.0.7
django-extensions==2.1.0

# breaking django-1.10:
# https://github.com/django-polymorphic/django-polymorphic/issues/225 - fixed
# https://github.com/makinacorpus/django-geojson/issues/82 - resolved in 2.9.1
# https://djaodjin.com/blog/subfieldbase-deprecated-durationfield.blog.html - fixed
# https://github.com/gadventures/django-fsm-admin/issues/63 - fixed (overwrote mixins.py with get_field() )
# https://github.com/saxix/django-adminactions/pull/104 - fixed by using own PR
# https://github.com/sehmaschine/django-grappelli/issues/783 - fixed by 2.9.1

# Admin
# ------------------------------------------------
django-adminactions==1.5
# -e git+https://github.com/saxix/django-adminactions.git@florianm-develop#egg=django-adminactions
django-grappelli==2.11.1


# API
# ------------------------------------------------
djangorestframework==3.8.2
djangorestframework-gis==0.13
djangorestframework-csv
djangorestframework-yaml
djangorestframework-jsonp
djangorestframework-filters
django-url-filter
#dynamic-rest
rest-framework-latex
# -e git+https://github.com/florianm/rest-framework-latex.git#egg=rest-framework-latex
markdown
django-filter==1.1.0
django-rest-swagger
coreapi==2.3.3
coreapi-cli
pypandoc
graphene-django==2.0.0
xmltodict==0.11.0

# Auth
# ------------------------------------------------
django-allauth==0.35.0
https://static.dpaw.wa.gov.au/static/py/dpaw-utils/dist/dpaw-utils-0.3a15.tar.gz

# Background tasks
# ------------------------------------------------
django-background-tasks


# DevOps
# ------------------------------------------------
#Fabric==2.2.1
Fabric3==1.14.post1
honcho
gunicorn==19.9.0

# Email backends for Mailgun, Postmark, SendGrid and more
# -------------------------------------------------------
django-anymail==1.2.1


# Configuration
# ------------------------------------------------
django-confy
django-environ==0.4.4
unipath


# Docs
# ------------------------------------------------
Sphinx>=1.6.1
flake8
pydot
sphinx_rtd_theme
pygraphviz


# Forms
# ------------------------------------------------
django-braces==1.12.0
django-crispy-forms==1.7.2
django-floppyforms==1.7.0
django-easy-select2==1.4.0
django_select2==6.0.2
django-bootstrap4
django-ajax-selects
django-map-widgets==0.1.9


# Models
# ------------------------------------------------
django-polymorphic==2.0.2
django-model-utils==3.1.1
django-reversion==2.0.13
django-fsm==2.6.0
django-fsm-log==1.5.0
django-fsm-admin==1.2.4
django-extra-fields==1.0.0
django-synctool
django-phonenumber-field==2.0.0
django-mptt


# Views
# ------------------------------------------------
django-tables2==1.21.2

# Images
# ------------------------------------------------
Pillow==5.1.0


# Python-PostgreSQL Database Adapter
# ------------------------------------------------
psycopg2==2.7.3.2

# Unicode slugification
# ------------------------------------------------
unicode-slugify==0.1.3
django-autoslug==1.9.3


# Time zones support
# ------------------------------------------------
pytz

# Templates
# ------------------------------------------------
django-settings-export

# Redis support
# ------------------------------------------------
django-redis==4.9.0
redis==2.10.6


# Map
# ------------------------------------------------
django-geojson>=2.9.1
django-leaflet==0.23.0


# Static and Media Storage
# ------------------------------------------------
whitenoise==3.3.0
django_compressor==2.1.1
boto==2.47.0
django-storages-redux==1.3.2
dj-static


# Profiling
# ------------------------------------------------
django-silk